+++ src/main/java/com/cliniconline/platform/model/dao/impl/DoctorDaoImpl.java
@@ -119,7 +119,7 @@ public class DoctorDaoImpl implements DoctorDao {
         doctor.setFirstName(rs.getString("f_name"));
         doctor.setLastName(rs.getString("l_name"));
         doctor.setPassword(rs.getInt("password"));
-        doctor.setPhoneNumber(rs.getInt("phone"));
+        doctor.setPhoneNumber(rs.getLong("phone"));
         doctor.setAddress(rs.getString("address"));
         doctor.setDOB(rs.getDate("dob"));
